Fix AFIR typo in locations module
AFIR is spelled out completely in the OCPI 2.3.0 specification as Alternative Fuel Infrasturcture Regulation. Note the typo in Infrasturcture. I corrected this typo, but this PR needs to be applied to all 2.3.0 related branches, but I'm unsure of the workflow. This branch seemed to have the most recent changes so I used this one, but happy to create one for other branches as well.